refactor: centralize types and pricing, fix hard-coded metrics
- Centralize JobStatus, JobSummary types in shared-types package
- Consolidate pricing data from evaluation-engine and web into shared-types
- Replace static metric lists with dynamic API calls to MetricRegistry
- Fix missing resultSnippet fields in test objects
- Create declaration files for proper TypeScript compilation
- Clean up circular dependencies in pricing imports
