+++ Dockerfile
@@ -22,8 +22,6 @@ RUN corepack enable
 COPY --from=builder /app/apps/web/dist            ./public
 COPY --from=builder /app/apps/api/dist            ./apps/api/dist
 COPY --from=builder /app/packages/evaluation-engine/dist        ./packages/evaluation-engine/dist
-COPY --from=builder /app/packages/evaluator/dist ./packages/evaluator/dist
-
 # ── package-manager metadata & production deps ───────────────────────────────
 COPY --from=builder /app/package.json \
                      /app/pnpm-lock.yaml \
@@ -31,10 +29,6 @@ COPY --from=builder /app/package.json \
                      ./
 COPY --from=builder /app/apps/api/package.json   ./apps/api/package.json
 COPY --from=builder /app/packages/evaluation-engine/package.json ./packages/evaluation-engine/package.json
-COPY --from=builder /app/packages/evaluator/package.json ./packages/evaluator/package.json
-COPY --from=builder /app/packages/test-cases/package.json ./packages/test-cases/package.json
-COPY --from=builder /app/packages/test-cases/src ./packages/test-cases/src
-
 # install all dependencies for CI/dev (including devDependencies for lint/test)
 RUN pnpm install --frozen-lockfile
