+++ apps/api/package.json
@@ -16,20 +16,17 @@
   "dependencies": {
     "@google/generative-ai": "^0.24.1",
     "@prompt-lab/evaluation-engine": "workspace:*",
-    "@prompt-lab/evaluator": "workspace:*",
     "compression": "^1.8.0",
     "dotenv": "^16.4.5",
     "express": "^4.21.2",
     "express-rate-limit": "^7.5.1",
     "openai": "^4.104.0",
-    "p-limit": "^5.0.0",
     "zod": "^3.22.2"
   },
   "devDependencies": {
     "@types/compression": "^1.8.1",
     "@types/express": "^4.17.21",
     "get-port": "^7.1.0",
-    "supertest": "^7.1.1",
     "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^3.2.4"
   }
