+++ apps/web/src/utils/tokenCounter.ts
@@ -1,23 +1,5 @@
-// Import pricing directly to avoid module initialization issues
-const PRICING = {
-  openai: {
-    'gpt-4.1-nano': { input: 0.0001, output: 0.0004 },
-    'gpt-4.1-mini': { input: 0.0004, output: 0.0016 },
-    'gpt-4.1': { input: 0.002, output: 0.008 },
-    'gpt-4o-mini': { input: 0.00015, output: 0.0006 },
-    'gpt-4o': { input: 0.0025, output: 0.01 },
-    'gpt-4-turbo': { input: 0.01, output: 0.03 },
-  },
-  gemini: {
-    'gemini-2.5-flash': { input: 0, output: 0 },
-    'gemini-1.5-pro': { input: 0.00125, output: 0.005 },
-  },
-  anthropic: {
-    'claude-3-5-haiku-20241022': { input: 0.0008, output: 0.004 },
-    'claude-3-5-sonnet-20241022': { input: 0.003, output: 0.015 },
-    'claude-3-opus-20240229': { input: 0.015, output: 0.075 },
-  },
-} as const;
+// Import pricing from shared-types
+import { PRICING } from '@prompt-lab/shared-types';
 
 // Model to encoding map for common OpenAI models
 const MODEL_ENCODINGS = {
